[FIX] Not displaying proper mode when using numeric input
Error description: formerly, when you would enter 1 (corresponding to encode) it would display "in 1 mode" instead of "in encode mode" ー now, it will display "in encode mode" instead.

New features testing: added the file_tools.py function library I wrote in preparation for the creation of the installer and updater. As is now, you can just download from GitHub, but there still needs to be an installer that will check for unmet dependencies like whether or not pyclip is installed or if git is installed (git is needed in order to update the program by doing a simple git clone)